All changes needed for the geode migration

+++ src/lists/views/PatchesListView.hpp
@@ -9,7 +9,7 @@ struct PatchesListView : public CustomListView {
 protected:
     virtual void setupList() override;
     virtual TableViewCell* getListCell(const char* key) override;
-    virtual void loadCell(TableViewCell* cell, const size_t index) override;
+    virtual void loadCell(TableViewCell* cell, const int index) override;
 private:
     size_t m_renderIndex;
 };
